requirements.txt
# Third-party libs and Django apps used by namesdbsite

                                   # LICENSE
django
djangorestframework
drf-yasg>=1.20.0,<1.21             # BSD      y
gunicorn>=20.1.0,<20.2             # MIT

git+https://github.com/IvanSaldoZ/django-bootstrap-pagination.git
git+https://github.com/denshoproject/densho-elastictools.git
